Move importer command to CLI
