+++ packages/importer/src/index.ts
@@ -8,3 +8,7 @@ export type ParseOptions = {
   noheader?: boolean;
   callback: (lines: string[][], columns: string[] | undefined, count: number) => Promise<boolean | void>;
 };
+
+export { parseFile as parseCSVFile, parseStream as parseCSVStream } from './csv';
+export { createProcessor } from './processor';
+export type { CompareSchemaResult, TableInfo } from './processor';
