+++ packages/importer/src/utils.test.ts
@@ -1,19 +1,6 @@
 import { describe, expect, test } from 'vitest';
 import { castType, guessType, guessTypes, normalizeColumnName, parseArray, parseRow, splitCommas } from './utils';
 
-describe('splitCommas', () => {
-  test('returns undefined for falsy values', () => {
-    expect(splitCommas(null)).toBeUndefined();
-    expect(splitCommas('')).toBeUndefined();
-    expect(splitCommas(false)).toBeUndefined();
-    expect(splitCommas(undefined)).toBeUndefined();
-  });
-
-  test('returns an array with the comma separated values', () => {
-    expect(splitCommas('a,b,c')).toEqual(['a', 'b', 'c']);
-  });
-});
-
 describe('normalizeColumnName', () => {
   test('transliterates, removes whitespaces and camel cases and keeps dots', () => {
     expect(normalizeColumnName(' 你好. buenos días ')).toBe('niHao.buenosDias');
